move examples out of pipelines directory
