+++ .pipelex/inference/backends.toml
@@ -27,7 +27,7 @@ enabled = false
 api_key = "${OPENAI_API_KEY}"
 
 [anthropic]
-enabled = false
+enabled = true
 api_key = "${ANTHROPIC_API_KEY}"
 claude_4_tokens_limit = 8192
 
@@ -55,4 +55,3 @@ api_key = "${FAL_API_KEY}"
 
 [internal] # software-only backend, runs internally, without AI
 enabled = true
-
