+++ .pipelex/inference/routing_profiles.toml
@@ -30,7 +30,7 @@ default = "pipelex_inference"
 # Pattern matching: "model-pattern" = "backend-name"
 "gpt-*" = "pipelex_inference"
 # "claude-*" = "pipelex_inference"
-"claude-*" = "bedrock"
+"claude-*" = "anthropic"
 "grok-*" = "pipelex_inference"
 "gemini-*" = "pipelex_inference"
 "*-sdxl" = "fal"
