+++ examples/extract_generic/__init__.py
