+++ examples/extract_generic/extract_generic.plx
