+++ examples/extract_proof_of_purchase/__init__.py
