+++ examples/extract_proof_of_purchase/extract_proof_of_purchase.plx
