+++ examples/extract_proof_of_purchase/models.py
