+++ examples/invoice_extractor/invoice_extractor.py
@@ -7,7 +7,7 @@ from pipelex.hub import get_pipeline_tracker, get_report_delegate
 from pipelex.pipelex import Pipelex
 from pipelex.pipeline.execute import execute_pipeline
 
-from pipelex_libraries.pipelines.examples.invoice_extractor.invoice import Invoice
+from examples.invoice_extractor.invoice import Invoice
 
 SAMPLE_NAME = "invoice_extractor"
 PDF_URL = "assets/invoice_extractor/invoice_1.pdf"
