+++ examples/simple_ocr.py
@@ -35,4 +35,4 @@ pretty_print(page_content_list)
 output_dir = get_results_dir_path(sample_name=SAMPLE_NAME)
 for page_index, page_content in enumerate(page_content_list.items):
     directory_for_page = f"{output_dir}/page_{page_index}"
-    page_content.save_to_directory(directory=directory_for_page)
+    # page_content.save_to_directory(directory=directory_for_page)
