+++ examples/wip/discord_newsletter/__init__.py
