--- examples/wip/photopposite.plx
@@ -1,82 +1,0 @@
-domain = "wip_domain"
-description = "Analyze a photo's most important feature and generate an image depicting its opposite"
-
-[concept]
-Photo = "A photograph"
-
-[concept.FeatureAnalysis]
-description = "Analysis of a photo's visual content and key features"
-
-[concept.FeatureAnalysis.structure]
-dominant_feature = { type = "text", description = "The most important or dominant feature in the image", required = true }
-visual_elements = { type = "text", description = "Description of key visual elements present", required = true }
-composition = "Analysis of the image composition"
-color_palette = "Description of the main colors in the image"
-mood_atmosphere = "The overall mood or atmosphere of the image"
-
-[concept.OppositeConcept]
-description = "Conceptual description of the opposite of a photo's most important feature"
-
-[concept.OppositeConcept.structure]
-opposite_feature = { type = "text", description = "The conceptual opposite of the dominant feature", required = true }
-visual_description = { type = "text", description = "Visual description of what the opposite would look like", required = true }
-reasoning = "Explanation of why this is considered the opposite"
-key_differences = "Key differences from the original"
-
-[concept.ImagePrompt]
-description = "Detailed prompt for generating an image"
-
-[pipe.gen_photopposite]
-type = "PipeSequence"
-description = "Analyze photo and generate its opposite"
-inputs = { photo = "Photo" }
-output = "Photo"
-steps = [
-    { pipe = "analyze_features", result = "feature_analysis" },
-    { pipe = "conceptualize_opposite", result = "opposite_concept" },
-    { pipe = "create_prompt", result = "image_prompt" },
-    { pipe = "generate_opposite", result = "opposite_photo" },
-]
-
-[pipe.analyze_features]
-type = "PipeLLM"
-description = "Analyze photo content and identify key features"
-inputs = { photo = "Photo" }
-output = "FeatureAnalysis"
-prompt = """Analyze this photo and identify its visual content and key features. Focus on identifying the single most important or dominant feature that defines this image.
-
-@photo
-
-Be concise.
-"""
-
-[pipe.conceptualize_opposite]
-type = "PipeLLM"
-description = "Determine the most important feature and conceptualize its opposite"
-inputs = { feature_analysis = "FeatureAnalysis" }
-output = "OppositeConcept"
-prompt = """Based on this feature analysis, determine the single most important feature and conceptualize its complete opposite. Describe what the opposite would look like in visual terms.
-
-@feature_analysis
-
-Be concise.
-"""
-
-[pipe.create_prompt]
-type = "PipeLLM"
-description = "Create detailed prompt for opposite photo"
-inputs = { opposite_concept = "OppositeConcept" }
-output = "ImagePrompt"
-prompt = """Create a concise image generation prompt that will produce a photo depicting this opposite concept. The prompt should be specific, visual, and suitable for photorealistic image generation.
-
-@opposite_concept
-
-Be concise.
-"""
-
-[pipe.generate_opposite]
-type = "PipeImgGen"
-description = "Generate the opposite photo"
-inputs = { image_prompt = "ImagePrompt" }
-output = "Photo"
-
