+++ quick_start/pipelines/example_1.plx
