+++ quick_start/pipelines/hello_world.plx
