+++ quick_start/pipelines/summarize.plx
