+++ quick_start/pipelines/summarize_struct.py
